Use (fixed-point) precision of NumPy array when writing DX file
Again, this is a workaround for an issue in the PyMol DX parser that fails to
parse tiny numbers (i.e. with an exponent lower than -100).  If it seems
inconsistent with requiring "type double", it is because the DX parser uses a
"float" to read the data:

static int read_dx_data(void *v, int set, float *datablock,
                         float *colorblock) {
  ...
  float grid;
  ...

Observed with PyMol up to 2.3.0.