Correct behavior for resample_factor
The reason tests were failing for the previous commit was that
resample_factor was depending on "nearest" interpolation to correct
for not altering the range where the edge points were calculated.
By correctly calculating an even distribution of edge points,
interpolation at "midpoints" no longer occurs outside the defined grid.